AoCElfTools

Documentation for AoCElfTools.

AoCElfTools.parse_input

AoCElfTools.parse_solve

AoCElfTools.samplepath

AoCElfTools.solve

AoCElfTools.userpath

AoCElfTools.parse_input — Method

parse_input(i::Int) -> (IO -> puzzle input)

Returns a function that parses Day i's puzzle input into a data structure that can then be used to find a solution.

Designed to be passed as a first argument to open.

Examples

julia> input = samplepath(1);

julia> println(read(input, String))
1000
2000
3000

4000

5000
6000

7000
8000
9000

10000

julia> open(parse_input(1), input)
5-element Vector{Int64}:
  6000
  4000
 11000
 24000
 10000

AoCElfTools.parse_solve — Method

parse_solve(i::Int) -> (IO -> solution)

Returns a function that combines the parsing and solution steps for Day i.

Designed to be passed as a first argument to open.

Examples

julia> input = samplepath(1);

julia> println(read(input, String))
1000
2000
3000

4000

5000
6000

7000
8000
9000

10000

julia> open(parse_solve(1), input)
(24000, 45000)

AoCElfTools.samplepath — Method

samplepath(i::Int) -> String

Return the path to the sample input file for Day i.

Examples

julia> samplepath(1)
"test/data/sample/01.txt"

julia> samplepath(14)
"test/data/sample/14.txt"

AoCElfTools.solve — Method

solve(i::Int) -> (puzzle input -> solution)

Returns a function that receive's Day i's puzzle input and returns a solution as a Tuple containing the two answers.

Examples

julia> input = open(parse_input(1), samplepath(1))
5-element Vector{Int64}:
  6000
  4000
 11000
 24000
 10000

julia> solve(1)(input)
(24000, 45000)

AoCElfTools.userpath — Method

userpath(user, i::Int) -> String

Return the path to user's input file for Day i.

Examples

julia> userpath("jbshannon", 1)
"test/data/jbshannon/01.txt"

julia> userpath("jbshannon", 13)
"test/data/jbshannon/13.txt"